+++ packages/engine/Source/DataSources/ModelGraphics.js
@@ -26,7 +26,7 @@ function createArticulationStagePropertyBag(value) {
  * @property {Property | boolean} [show=true] A boolean Property specifying the visibility of the model.
  * @property {Property | string | Resource} [uri] A string or Resource Property specifying the URI of the glTF asset.
  * @property {Property | number} [scale=1.0] A numeric Property specifying a uniform linear scale.
- * @property {Property | boolean} [allowVerticalExaggeration=true] A boolean Property specifying to allow participation in scene Vertical Exaggeration.
+ * @property {Property | boolean} [enableVerticalExaggeration=true] A boolean Property specifying if the model is exaggerated along the ellipsoid normal when {@link Scene.verticalExaggeration} is set to a value other than <code>1.0</code>.
  * @property {Property | number} [minimumPixelSize=0.0] A numeric Property specifying the approximate minimum pixel size of the model regardless of zoom.
  * @property {Property | number} [maximumScale] The maximum scale size of a model. An upper limit for minimumPixelSize.
  * @property {Property | boolean} [incrementallyLoadTextures=true] Determine if textures may continue to stream in after the model is loaded.
@@ -73,8 +73,8 @@ function ModelGraphics(options) {
   this._scaleSubscription = undefined;
   this._verticalExaggerationOn = undefined;
   this._verticalExaggerationOnSubscription = undefined;
-  this._allowVerticalExaggeration = undefined;
-  this._allowVerticalExaggerationSubscription = undefined;
+  this._enableVerticalExaggeration = undefined;
+  this._enableVerticalExaggerationSubscription = undefined;
   this._minimumPixelSize = undefined;
   this._minimumPixelSizeSubscription = undefined;
   this._maximumScale = undefined;
@@ -156,14 +156,13 @@ Object.defineProperties(ModelGraphics.prototype, {
   scale: createPropertyDescriptor("scale"),
 
   /**
-   * Gets or sets the boolean Property specifying the model to
-   * participate in scene Vertical Exaggeration.
+   * Gets or sets the boolean Property specifying if the model is exaggerated along the ellipsoid normal when {@link Scene.verticalExaggeration} is set to a value other than <code>1.0</code>.
    * @memberof ModelGraphics.prototype
    * @type {Property|undefined}
-   * @default false
+   * @default true
    */
-  allowVerticalExaggeration: createPropertyDescriptor(
-    "allowVerticalExaggeration"
+  enableVerticalExaggeration: createPropertyDescriptor(
+    "enableVerticalExaggeration"
   ),
 
   /**
@@ -349,7 +348,7 @@ ModelGraphics.prototype.clone = function (result) {
   result.show = this.show;
   result.uri = this.uri;
   result.scale = this.scale;
-  result.allowVerticalExaggeration = this.allowVerticalExaggeration;
+  result.enableVerticalExaggeration = this.enableVerticalExaggeration;
   result.minimumPixelSize = this.minimumPixelSize;
   result.maximumScale = this.maximumScale;
   result.incrementallyLoadTextures = this.incrementallyLoadTextures;
@@ -387,9 +386,9 @@ ModelGraphics.prototype.merge = function (source) {
   this.show = defaultValue(this.show, source.show);
   this.uri = defaultValue(this.uri, source.uri);
   this.scale = defaultValue(this.scale, source.scale);
-  this.allowVerticalExaggeration = defaultValue(
-    this.allowVerticalExaggeration,
-    source.allowVerticalExaggeration
+  this.enableVerticalExaggeration = defaultValue(
+    this.enableVerticalExaggeration,
+    source.enableVerticalExaggeration
   );
   this.minimumPixelSize = defaultValue(
     this.minimumPixelSize,